add navigation and refactor
