+++ app/build.gradle
@@ -60,6 +60,10 @@ dependencies {
     // Coil
     implementation 'io.coil-kt:coil:2.2.2'
 
+    // Navigation
+    implementation "androidx.navigation:navigation-fragment-ktx:2.5.3"
+    implementation "androidx.navigation:navigation-ui-ktx:2.5.3"
+
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.6.0'
     implementation 'com.google.android.material:material:1.8.0'
