+++ app/src/main/java/com/example/movietop/di/BindsModule.kt
@@ -1,12 +1,12 @@
 package com.example.movietop.di
 
-import com.example.movietop.data.movies.MoviesRepositoryImpl
-import com.example.movietop.domain.movies.MoviesRepository
+import com.example.movietop.data.movies.list.MoviesListRepositoryImpl
+import com.example.movietop.domain.movies.list.MoviesListRepository
 import dagger.Binds
 import dagger.Module
 
 @Module
 interface BindsModule {
     @Binds
-    fun bindMoviesListRepository(implementation: MoviesRepositoryImpl): MoviesRepository
+    fun bindMoviesListRepository(implementation: MoviesListRepositoryImpl): MoviesListRepository
 }
