--- app/src/main/java/com/example/movietop/domain/movies/MoviesRepository.kt
@@ -1,8 +1,0 @@
-package com.example.movietop.domain.movies
-
-import androidx.paging.PagingData
-import kotlinx.coroutines.flow.Flow
-
-interface MoviesRepository {
-    fun getPagedMovies(): Flow<PagingData<Film>>
-}
