+++ app/src/main/java/com/example/movietop/domain/movies/list/MoviesListRepository.kt
@@ -1,0 +1,9 @@
+package com.example.movietop.domain.movies.list
+
+import androidx.paging.PagingData
+import com.example.movietop.domain.movies.Film
+import kotlinx.coroutines.flow.Flow
+
+interface MoviesListRepository {
+    fun getPagedMovies(): Flow<PagingData<Film>>
+}
